data/requirements.txt
Flask==3.1.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
python-dotenv==1.0.1
waitress==3.0.0
Werkzeug==3.1.3
requests==2.32.3
matplotlib==3.10.3
numpy==2.2.3
geojson==3.2.0
shapely==2.1.0
sentinelhub==3.11.1
